Fix AAE package name.
