+++ aae/src/test/java/com/example/android/uamp/aae/ExampleUnitTest.java
@@ -14,30 +14,20 @@
  * limitations under the License.
  */
 
-package com.example.aae;
-
-import android.content.Context;
+package com.example.android.uamp.aae;
 
 import org.junit.Test;
-import org.junit.runner.RunWith;
 
-import androidx.test.InstrumentationRegistry;
-import androidx.test.runner.AndroidJUnit4;
-
-import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.*;
 
 /**
- * Instrumented test, which will execute on an Android device.
+ * Example local unit test, which will execute on the development machine (host).
  *
  * @see <a href="http://d.android.com/tools/testing">Testing documentation</a>
  */
-@RunWith(AndroidJUnit4.class)
-public class ExampleInstrumentedTest {
+public class ExampleUnitTest {
     @Test
-    public void useAppContext() {
-        // Context of the app under test.
-        Context appContext = InstrumentationRegistry.getTargetContext();
-
-        assertEquals("com.example.aae", appContext.getPackageName());
+    public void addition_isCorrect() {
+        assertEquals(4, 2 + 2);
     }
 }
